requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.7.2
psycopg2-binary==2.9.1
sqlalchemy==1.4.23
pydantic